% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmkl.R
\name{getker}
\alias{getker}
\title{Get kernels for data}
\usage{
getker(x)
}
\arguments{
\item{x}{input matrix}
}
\value{
the kernels
}
\description{
Get kernels with only one input matrix
}
\examples{
x <- matrix(rnorm(12), 3, 4)
kk <- getker(x)
}
