% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmkl.R
\name{getkertest}
\alias{getkertest}
\title{Get kernels for data}
\usage{
getkertest(xtr, xte)
}
\arguments{
\item{xtr}{training matrix}

\item{xte}{test matrx}
}
\value{
the kernels
}
\description{
Get kernels with two input matrices
}
\examples{
x <- matrix(rnorm(9), 3, 3)
y <- matrix(rnorm(9), 3, 3)
kk <- getkertest(x, y)
}
