% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmkl.R
\name{predict.cv_glmmkl}
\alias{predict.cv_glmmkl}
\title{Prediction of MKL models}
\usage{
\method{predict}{cv_glmmkl}(object, x, ...)
}
\arguments{
\item{object}{MKL model}

\item{x}{the new kernels}

\item{...}{\dots}
}
\value{
the predicted values
}
\description{
Prediction for MKL models
}
\examples{
x <- matrix(rnorm(200), 100, 2)
y <- x \%*\% rnorm(2, 1) + rnorm(100)
x1 <- x[1:80, ]
x2 <- x[81:100, ]
y1 <- y[1:80]
res <- cv_glmmkl(x1, y1)
pres <- predict(res, getkertest(x1, x2))
}
